Move `getSignupStepsSiteTopic()` selector under `signup/steps/site-topic` state tree directory.
