--- client/state/selectors/get-signup-steps-site-topic.js
@@ -1,8 +1,0 @@
-/** @format */
-
-/**
- * External dependencies
- */
-import { get } from 'lodash';
-
-export default state => get( state, 'signup.steps.siteTopic', null );
