+++ client/state/signup/steps/site-topic/selectors.js
@@ -1,0 +1,8 @@
+/** @format */
+
+/**
+ * External dependencies
+ */
+import { get } from 'lodash';
+
+export const getSignupStepsSiteTopic = state => get( state, 'signup.steps.siteTopic', null );
